requirements.txt
pandas
numpy
faiss-cpu
sentence-transformers
openai
rich
python-dotenv
langchain
langchain-openai
langchain-community
langchain-huggingface
fastapi
uvicorn
jinja2
python-multipart